Create y delete reposts funciona. No puedo borrar posts o reposts que no son míos. No puedo repostear el mismo post dos veces. Si borro un repost, no se borran otros reposts ni el post original. si borro un post, se borran los reposts

+++ repository/queries/queries_reposts.py
@@ -1,13 +1,13 @@
 """
 Queries for creating and deleting reposts
 """
-
-from typing import List
 from sqlalchemy import Delete
 from sqlalchemy.exc import IntegrityError
 
 # pylint: disable=C0114, W0401, W0614, E0602, E0401
 from repository.queries.common_setup import *
+
+from repository.queries.queries_global import is_public
 
 # pylint: disable=C0114, W0401, W0614, E0401
 from repository.errors import (
@@ -19,6 +19,7 @@ from repository.errors import (
     PostNotFound,
     DatabaseError,
     UserWithouPermission,
+    RepostAlreadyMade
 )
 
 # pylint: disable=C0114, W0401, W0614, E0401
@@ -32,30 +33,40 @@ def create_repost(post_id: int, user_reposter_id: int):
     Create a repost of a post
     """
     try:
-        post_and_content = (session.query(Content)
-                            .join(Post, Post.content_id == Content.content_id)
+        post = (session.query(Post)
                             .filter(Post.post_id == post_id)
                             .first())
         
         #original post to repost not found
-        if post_and_content is None:
+        if post is None:
             raise PostNotFound()
         
         #user wants to repost a post --> not allowed
-        if post_and_content.user_creator_id != post_and_content.user_poster_id:
+        if post.user_creator_id != post.user_poster_id:
             raise UserWithouPermission()
         
         #user wants to repost from a private account --> not allowed
-        if not user_is_public(post_and_content.user_creator_id):
+        if not is_public(post.user_creator_id):
             raise UserWithouPermission()
+
+        repost_check = (session.query(Post)
+                            .filter(Post.user_poster_id == user_reposter_id)
+                            .filter(Post.user_creator_id == post.user_creator_id)
+                            .filter(Post.content_id == post.content_id)
+                            .first())
         
-        post = Post(
+        #user wants to repost the same post twice
+        if repost_check is not None:
+            raise RepostAlreadyMade()
+        
+        repost = Post(
             user_poster_id = user_reposter_id,
-            user_creator_id = post_and_content.user_creator_id,
-            content_id = post_and_content.content_id,
+            user_creator_id = post.user_creator_id,
+            content_id = post.content_id,
         )
 
-        return
+        session.add(repost)
+        session.commit()
     except IntegrityError as error:
         session.rollback()
         raise DatabaseError from error
@@ -63,14 +74,22 @@ def create_repost(post_id: int, user_reposter_id: int):
 
 # ----- DELETE ------
 
-def delete_repost(post_id):
+def delete_repost(post_id, user_id):
     """
     Deletes the repost with that id, if found
     """
     try:
-        repost = session.query(Post).filter(Post.id == post_id).first()
+        repost = session.query(Post).filter(Post.post_id == post_id).first()
         if repost is None:
             raise PostNotFound()
+        
+        #trying to delete a post that was not made by this user
+        if repost.user_poster_id != user_id: 
+            raise UserWithouPermission()
+        
+        #trying to delete a post --> wrong endpoint
+        if repost.user_creator_id == user_id: 
+            raise UserWithouPermission()
         
         session.delete(repost)
         session.commit()
